[change] refactoring timeline task component.
