+++ src/app/javascripts/renderer/components/taskbord/timeline-task.jsx
@@ -32,36 +32,19 @@ const propTypes = {
   connectDragSource: PropTypes.func.isRequired
 }
 
-const TimelineTask = class TvTask extends React.Component {
+const timelineTask = class TimelineTask extends React.Component {
   render() {
     const { isDragging, connectDragSource, text } = this.props
-    let index = 1;
-    let tasks = [];
-    let top = 50
 
-    this.props.taskList.document.nodes.map((block) => {
-      if (block.text != "") {
-        let height = 49 * block.data.get("requiredTime", 60) / 60
-        let style = {
-          top: top.toString() + 'px',
-          height: height.toString() + 'px'
-        };
-        tasks.push(<div key={block.key} className={block.data.get("done", false) ? "task done" : "task"} style={style}><span>{block.text}</span></div>);
-        top = top + height + 1
-      }
-    })
-
-    if (tasks) {
-      return connectDragSource(
-        <td className="tv-task">
-          {tasks}
-          <div className="nowmarker"></div>
-        </td>
-      );
-    } else {
-      return null;
-    }
+    return connectDragSource(
+      <div
+        key={this.props.block.key}
+        className={this.props.block.data.get("done", false) ? "task done" : "task"}
+        style={this.props.style}>
+        <span>{this.props.block.text}</span>
+      </div>
+    )
   }
 }
 
-module.exports = DragSource(Types.CARD, taskSource, collect)(TimelineTask);
+module.exports = DragSource(Types.CARD, taskSource, collect)(timelineTask);
